Moved SectionTitle to /ui
