+++ src/components/portfolio/Experience.tsx
@@ -1,5 +1,5 @@
 "use client"
-import SectionTitle from "@/components/portfolio/SectionTitle";
+import SectionTitle from "@/components/ui/SectionTitle";
 import experiences from "@/data/experience";
 import { Timeline, Avatar } from '@mantine/core';
